requirements.txt
pystray==0.19.5
Pillow==10.1.0
watchdog==3.0.0
plyer==2.1.0
PyQt5==5.15.9
PyQt5-sip==12.13.0
